+++ packages/x-components/src/x-modules/next-queries/store/actions/fetch-and-save-next-queries.action.ts
@@ -1,6 +1,6 @@
 import { NextQuery } from '@empathyco/x-types';
 // eslint-disable-next-line max-len
-import { createFetchAndSaveAction } from '../../../../store/utils/fetch-and-save-action.helpers';
+import { createFetchAndSaveAction } from '../../../../store/utils/fetch-and-save-action.utils';
 import { NextQueriesActionContext } from '../types';
 
 const { fetchAndSave, cancelPrevious } = createFetchAndSaveAction<
