+++ packages/x-components/src/x-modules/popular-searches/store/actions/fetch-and-save-suggestions.action.ts
@@ -1,6 +1,6 @@
 import { Suggestion } from '@empathyco/x-types';
 // eslint-disable-next-line max-len
-import { createFetchAndSaveAction } from '../../../../store/utils/fetch-and-save-action.helpers';
+import { createFetchAndSaveAction } from '../../../../store/utils/fetch-and-save-action.utils';
 import { PopularSearchesActionContext } from '../types';
 
 const { fetchAndSave, cancelPrevious } = createFetchAndSaveAction<
