+++ packages/x-components/src/x-modules/popular-searches/store/types.ts
@@ -1,7 +1,7 @@
 import { SuggestionsRequest } from '@empathyco/x-adapter';
 import { HistoryQuery, Suggestion } from '@empathyco/x-types';
 import { XActionContext, XStoreModule } from '../../../store';
-import { StatusMutations, StatusState } from '../../../store/utils/status.helpers';
+import { StatusMutations, StatusState } from '../../../store/utils/status-store.utils';
 import { PopularSearchesConfig } from '../config.types';
 import { Dictionary } from '../../../utils/types';
