+++ packages/x-components/src/x-modules/query-suggestions/store/actions/fetch-and-save-suggestions.action.ts
@@ -1,6 +1,6 @@
 import { Suggestion } from '@empathyco/x-types';
 // eslint-disable-next-line max-len
-import { createFetchAndSaveAction } from '../../../../store/utils/fetch-and-save-action.helpers';
+import { createFetchAndSaveAction } from '../../../../store/utils/fetch-and-save-action.utils';
 import { QuerySuggestionsActionContext } from '../types';
 
 const { fetchAndSave, cancelPrevious } = createFetchAndSaveAction<
