+++ packages/x-components/src/x-modules/recommendations/store/module.ts
@@ -1,4 +1,4 @@
-import { setStatus } from '../../../store/utils/status.helpers';
+import { setStatus } from '../../../store/utils/status-store.utils';
 import {
   cancelFetchAndSaveRecommendations,
   fetchAndSaveRecommendations
