+++ packages/x-components/src/x-modules/related-tags/store/actions/fetch-and-save-related-tags.action.ts
@@ -1,6 +1,6 @@
 import { RelatedTag } from '@empathyco/x-types';
 // eslint-disable-next-line max-len
-import { createFetchAndSaveAction } from '../../../../store/utils/fetch-and-save-action.helpers';
+import { createFetchAndSaveAction } from '../../../../store/utils/fetch-and-save-action.utils';
 import { RelatedTagsActionContext } from '../types';
 
 const { fetchAndSave, cancelPrevious } = createFetchAndSaveAction<
