+++ packages/x-components/src/x-modules/search/store/types.ts
@@ -11,7 +11,7 @@ import {
   Sort
 } from '@empathyco/x-types';
 import { XActionContext, XStoreModule } from '../../../store';
-import { StatusMutations, StatusState } from '../../../store/utils/status.helpers';
+import { StatusMutations, StatusState } from '../../../store/utils/status-store.utils';
 import { QueryOrigin } from '../../../types/query-origin';
 import { Dictionary } from '../../../utils/types';
 import { SearchConfig } from '../config.types';
